# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with authentication, roles and encryption.",
        "title": "Go API Dojo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "In bearer/dual mode, returns a token pair in JSON. In cookie-only mode, sets HttpOnly auth cookies and returns {\"authenticated\":true}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate a user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "2FA required — complete OTP flow",
                        "schema": {
                            "$ref": "#/definitions/handler.OTPRequiredResponse"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "account suspended or not verified",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes server-side refresh sessions. In cookie/dual mode, also clears auth cookies on the client.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke all active refresh tokens for the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required in cookie/dual mode",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "csrf validation failed",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/otp/resend": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Resend the OTP code by email",
                "parameters": [
                    {
                        "description": "Resend payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ResendRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/otp/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate a 2FA OTP code and return a token pair",
                "parameters": [
                    {
                        "description": "OTP payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.VerifyOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.TokenPair"
                        }
                    },
                    "400": {
                        "description": "invalid or expired code",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/password/new": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Set a new password using a reset token",
                "parameters": [
                    {
                        "description": "Reset payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/password/reset": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Send a password reset email",
                "parameters": [
                    {
                        "description": "Email payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SendPasswordResetRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid email",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "email already in use",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/token/refresh": {
            "post": {
                "description": "Accepts refresh token from JSON body or, in cookie/dual mode, from the refresh cookie. In cookie-only mode, rotated tokens are returned via HttpOnly cookies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Rotate a refresh token and return a new token pair",
                "parameters": [
                    {
                        "description": "Refresh token payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RefreshTokenRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Required in cookie/dual mode",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cookie mode: tokens are set in HttpOnly cookies",
                        "schema": {
                            "$ref": "#/definitions/handler.RefreshCookieModeResponse"
                        }
                    },
                    "401": {
                        "description": "invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "csrf validation failed",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify an email address using the token sent by email",
                "parameters": [
                    {
                        "description": "Verification payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.VerifyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify/resend": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Resend the account verification email",
                "parameters": [
                    {
                        "description": "Resend payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ResendRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/fault.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fault.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "a descriptive error message"
                }
            }
        },
...
        "handler.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 160,
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "Password1!"
                }
            }
        },
...
        "handler.OTPRequiredResponse": {
            "type": "object",
            "properties": {
                "otp_required": {
                    "type": "boolean",
                    "example": true
                },
                "user_id": {
                    "type": "string",
                    "example": "01966b0a-1234-7abc-def0-1234567890ab"
                }
            }
        },
...
        "handler.ResendRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "01966b0a-1234-7abc-def0-1234567890ab"
                }
            }
        },
...
        "handler.SendPasswordResetRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
...
        "handler.VerifyOTPRequest": {
            "type": "object",
            "required": [
                "code",
                "user_id"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "user_id": {
                    "type": "string",
                    "example": "01966b0a-1234-7abc-def0-1234567890ab"
                }
            }
        },
...
        "service.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refresh_token": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6..."
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}